Add files via upload
In this code, the accuracy on the test set is low. The reason for that is because the dataset was quite sparse and with a different preprocessing approach the accuracy could have improved. We wanted to try this using CNN as it was an interesting idea. The aim of the project was to not beat state of the art score but to simply use the concepts taught in the course.